% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseRootData.R
\name{analyseRootData}
\alias{analyseRootData}
\title{Calculates average of Root Measurements}
\usage{
analyseRootData(inputFile, controlTreatment, outputFile)
}
\description{
Calculates average of Root Measurements
}
