% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRootBarG.R
\name{createRootBarG}
\alias{createRootBarG}
\title{Plot Root Length Data}
\usage{
createRootBarG(inputFile)
}
\arguments{
\item{inputFile}{A string representing the path for the input file containing
root measurement data. Precondition: columns with data to be plotted must
include / character. Columns of gene lines must be named "geneLines"}
}
\value{
A bar plot of type list
}
\description{
A function that plots the root length paramters based on treatment and gene
Lines.
}
\examples{
# Example 1:
# Creating Bar plot of input data
\dontrun{
library(plyr)
input <- system.file("extdata", "inputFileFunc3.csv", package = "FastRPA")
filterData <- createRootBarG(input)
}

}
\references{
Steipe B., ABC project (.utility 4.07) A Bioinformatics Course:
Applied Bioinformatics
\href{http://steipe.biochemistry.utoronto.ca/abc/index.php/Bioinformatics_Main_Page}{Link}.
}
\author{
Erik Ensminger \email{erik.ensminger@mail.utoronto.ca}
}
