% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputFileFunc1}
\alias{inputFileFunc1}
\title{Root Length Measurements}
\format{
A matrix with columns:
\describe{
\item{Condition.1}{First condition, Column with root length measurements must be called 'Longest Shortest Path'}
}
}
\source{
Ensminger, Erik
}
\usage{
inputFileFunc1
}
\description{
A file containing the raw data of the root length measurements used as input for function filterData
}
\examples{
# Example 1:
# Access raw data made available with this package
inputRootLength <- system.file("extdata", "inputFileFunc1.csv", package = "TestingPackage")
# Read data
inputFileFunc1 <- read.csv(file = inputRootLength, header = TRUE)
dim(inputFileFunc1) # 30  4
# To see documentation for this dataset
?inputFileFunc1

}
\keyword{datasets}
