% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputFileFunc2}
\alias{inputFileFunc2}
\title{Input File of Function 2, analyseRootData}
\format{
A matrix with columns:
\describe{
\item{Condition.1}{First condition, first column contains geneLines and must be called 'geneLines'}
\item{Condition.2}{Second condition, subsequent columns must contain the name of the experimental root treatment}
\item{Condition.2}{Third condition, measurments of experimental root treatment must be in milimetres (mm)}
\item{Condition.3}{Fourth condition, one of the root treatment columns must be control treatment and must be listed as the last column in the csv file, which will be used for normalizing data}
}
}
\source{
Ensminger, Erik
}
\usage{
inputFileFunc2
}
\description{
The file inputFileFunc2 contains the filtered data outputted by the filterData function
This file can then be used as input for function analyseRootData
}
\examples{
# Example 1:
# Access raw data made available with this package
inputfilteredData <- system.file("extdata", "inputFileFunc2.csv", package = "FastRPA")
# Read data
inputFileFunc2 <- read.csv(file = inputfilteredData, header = TRUE)
dim(inputFileFunc2) # 30  4
# To see documentation for this dataset
?inputFileFunc2

}
\keyword{datasets}
