% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputFileFunc3}
\alias{inputFileFunc3}
\title{Input File of Function 3, createRootBarG}
\format{
A matrix with columns:
\describe{
\item{Condition.1}{First condition, first column contains geneLines and must be called 'geneLines'}
\item{Condition.2}{Second condition, the name of root measurements columns to be plotted must contain a \verb{\\} character}
\item{Condition.2}{Third condition, measurments of experimental root treatment must be in milimetres (mm)}
}
}
\source{
Ensminger, Erik
}
\usage{
inputFileFunc3
}
\description{
The file inputFileFunc3 contains the normalized root length data outputted by the analyseRootData function
This file can then be used as input for function createRootBarG
}
\examples{
# Example 1:
# Access raw data made available with this package
inputnormalizedData <- system.file("extdata", "inputFileFunc3.csv", package = "FastRPA")
# Read data
inputFileFunc3 <- read.csv(file = inputnormalizedData, header = TRUE)
dim(inputFileFunc3) # 30  4
# To see documentation for this dataset
?inputFileFunc3

}
\keyword{datasets}
