% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outputFileFunc1}
\alias{outputFileFunc1}
\title{Output File of Function 1}
\format{
A matrix with columns:
\describe{
\item{Condition.1}{First condition, first column contains geneLines and must be called 'geneLines'}
\item{Condition.2}{Second condition, subsequent columns must contain the name of the experimental root treatment}
\item{Condition.3}{Second condition, one of the root treatment columns must be control treatment, which will be used for normalizing data}
}
}
\source{
Ensminger, Erik
}
\usage{
outputFileFunc1
}
\description{
The file outputFileFunc1 contains the filtered data outputted by the filterData function
This file may also be used as the input for function 2, analyseRootData
}
\examples{
# Example 1:
# Access raw data made available with this package
outputfilteredData <- system.file("extdata", "inputFileFunc1.csv", package = "TestingPackage")
# Read data
outputFileFunc1 <- read.csv(file = outputfilteredData, header = TRUE)
dim(outputFileFunc1) # 30  4
# To see documentation for this dataset
?outputFileFunc1

}
\keyword{datasets}
