% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outputFileFunc2}
\alias{outputFileFunc2}
\title{Output File of Function 2, analyseRootData}
\format{
A matrix with columns:
\describe{
\item{Condition.1}{First condition, first column contains geneLines and must be called 'geneLines'}
\item{Condition.2}{Second condition, the name of root measurements columns to be plotted must contain a \verb{\\} character}
\item{Condition.2}{Third condition, measurments of experimental root treatment must be in milimetres (mm)}
}
}
\source{
Ensminger, Erik
}
\usage{
outputFileFunc2
}
\description{
The file outputFileFunc2 contains the normalized root length data outputted by the analyseRootData function
This file can then be used as input for function createRootBarG
}
\examples{
# Example 1:
# Access raw data made available with this package
outputnormalizedData <- system.file("extdata", "outputFileFunc2.csv", package = "FastRPA")
# Read data
outputFileFunc2 <- read.csv(file = outputnormalizedData, header = TRUE)
dim(outputFileFunc2) # 30  4
# To see documentation for this dataset
?outputFileFunc2

}
\keyword{datasets}
