% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseRootData.R
\name{analyseRootData}
\alias{analyseRootData}
\title{Calculates average of Root Measurements}
\usage{
analyseRootData(inputFile, controlTreatment, outputFile)
}
\arguments{
\item{inputFile}{A string representing the path for the .csv file containing Root Length Data}

\item{controlTreatment}{A string representing the control Treatment included in inputFile}

\item{outputFile}{A name of a files to which the calculated averages will be exported}
}
\value{
Output is data frame of data output file. In addition, data is exported to the outputFile
}
\description{
Given a user defined inputFile path, which contains the GeneLines names,
and the root length values of the control treatments, the function will
normalize the root length measurements. For this, the user must include
and specify which treatment in the inputFile is the control treatment.
}
\details{
After calculating the average root length of the control treatment for each
root line, and the average root length of each treatment, the treatments will
be normalized by the control treatment average.

The function will then output the new root output data to a user specified output
file
}
\examples{
\dontrun{
#Example 1:
# Normalizing root data and calculating average of root treatments.
#file inputFileFunc2, first column are the gene Lines, MEX, TEX and WES columns are 3 differnet root treatments
input <- system.file("extdata", "inputFileFunc2.csv", package = "FastRPA")
outputFile <- system.file("extdata", "outputFileFunc2.csv", package = "FastRPA")
analyseRoot <- analyseRootData(inputFile = input, controlTreatment = "WES",
                            outputFile = outputFile)
}

}
