% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{Filter Root length Measurements}
\usage{
filterData(inputFile, roi = c(), rootLines, rootTreatment, outputFile)
}
\arguments{
\item{inputFile}{A string representing the path for the input file containing
root measurement data}

\item{roi}{a vector containing the index of specified root measurements to be
exported to the outputFile}

\item{rootLines}{A string representing the root line the specified roi root
measurements parameters are.}

\item{rootTreatment}{A string representing the root treatment the specified
roi root root measurements parameters are.}

\item{outputFile}{A string representing the path for the output file to which
root measurements are added}
}
\value{
return is a dataframe which will also be exported to outputFile
}
\description{
A function that takes root measurements selected by the user and adds them to
a new output file specified by the user.
}
\examples{
# Example 1:
# Adding two selected root measurement to output file
\dontrun{
library(plyr)
input <- system.file("extdata", "inputFileFunc1.csv", package = "FastRPA")
roi <- c(2,1,5)
outputFile <- system.file("extdata", "outputFileFunc1.csv", package = "FastRPA")
filterData <- filterData(inputFile = input, roi = roi, rootLines = "ASER.12.32",
                       rootTreatment = "TEX", outputFile = outputFile)
}

}
\references{
Steipe B., ABC project (.utility 4.07) A Bioinformatics Course:
Applied Bioinformatics
\href{http://steipe.biochemistry.utoronto.ca/abc/index.php/Bioinformatics_Main_Page}{Link}.
}
\author{
Erik Ensminger \email{erik.ensminger@mail.utoronto.ca}
}
